Enforce UDP size limit
Summary:
In the current code, UDP payloads larger than 1500 bytes will be truncated to 1500 bytes, and Datagram payloads larger than 1500 bytes will be forwarded as is (potentially resulting in truncation by the OS).  Truncation is not the intended behavior for CONNECT-UDP.

This patch changes the behavior to drop payloads larger than 1500 bytes, in either direction.  (Datagram payloads smaller than 1500 bytes if they are larger than the interface MTU.)