Reduce log spam on connection shutdown
This change adds a function that identifies
errors that are known to be ignorable.  It
applies this check across the proxying
functions to reduce log spam when the proxy
is shut down, or an individual connection is
closed.

This avoids the need to check `s.closed`,
so the proxying logic can be moved from a
method to a function (see next PR).