Refactor virtual connection as a thin wrapper around net.Pipe
Summary:
The current virtual connection struct `proxiedConn` implements the client side of CONNECT-UDP.  It also contains a great deal of logic to pair reads and writes using a channel, implement deadlines using timers, etc.  Most of this functionality is already available in `masque.Proxy` or `net.Pipe`.

This change reorganizes conn.go and proxy.go to share their implementation of the CONNECT-UDP protocol (which is logically symmetrical for clients and servers), using `net.Pipe` to reverse the direction on the client side.  This reduces the amount of code required in this package and ensures symmetry between the client and server implementations.

The only significant behavior change here is that calls to `ReceiveDatagram(ctx)` no longer pass a nontrivial context, so they cannot time out.  This may result in one packet of additional buffering.

The main motivation for this change is to enable a sensible implementation of the Capsule Protocol.